fix some rebase issues
