+++ crates/bevy_core/Cargo.toml
@@ -19,5 +19,4 @@ bevy_tasks = { path = "../bevy_tasks", version = "0.8.0-dev" }
 bevy_utils = { path = "../bevy_utils", version = "0.8.0-dev" }
 
 # other
-crossbeam-channel = "0.5.0"
 bytemuck = "1.5"
