+++ crates/bevy_time/src/lib.rs
@@ -9,6 +9,10 @@ pub use stopwatch::*;
 pub use time::*;
 pub use timer::*;
 
+use bevy_ecs::system::{Local, Res, ResMut};
+use bevy_utils::{tracing::warn, Instant};
+use crossbeam_channel::{Receiver, Sender};
+
 pub mod prelude {
     //! The Bevy Time Prelude.
     #[doc(hidden)]
@@ -41,6 +45,36 @@ impl Plugin for TimePlugin {
     }
 }
 
-fn time_system(mut time: ResMut<Time>) {
-    time.update();
+/// channel resource used to receive time from render world
+pub struct TimeReceiver(pub Receiver<Instant>);
+/// channel resource used to send time from render world
+pub struct TimeSender(pub Sender<Instant>);
+
+/// create channels used for sending time between render world and app world
+pub fn create_time_channels() -> (TimeSender, TimeReceiver) {
+    // bound the channel to 2 since when pipelined the render phase can finish before
+    // the time system runs.
+    let (s, r) = crossbeam_channel::bounded::<Instant>(2);
+    (TimeSender(s), TimeReceiver(r))
+}
+
+/// The system used to update the time. If there is a render world the time is sent from
+/// there to this system through channels. Otherwise the time is updated in this system.]
+fn time_system(
+    mut time: ResMut<Time>,
+    time_recv: Option<Res<TimeReceiver>>,
+    mut has_received_time: Local<bool>,
+) {
+    if let Some(time_recv) = time_recv {
+        // TODO: delay checking channel on start by 2 frames when pipelined rendering
+        // is enabled. This is to make sure we always read the N-2 frame's time.
+        if let Ok(new_time) = time_recv.0.try_recv() {
+            time.update_with_instant(new_time);
+            *has_received_time = true;
+        } else if *has_received_time {
+            warn!("time_system did not receive time from render world! Calculations depending on the time may be incorrect.");
+        }
+    } else {
+        time.update();
+    }
 }
